chore: split token cache provider into more files
