+++ internal/serviceaccounttokens/cache/tokens.go
@@ -1,0 +1,86 @@
+// MIT License
+//
+// Copyright (c) 2024 Matheus Pimenta
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in all
+// copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+// SOFTWARE.
+
+package cacheserviceaccounttokens
+
+import (
+	"context"
+	"fmt"
+	"time"
+
+	"github.com/matheuscscp/gke-metadata-server/internal/serviceaccounts"
+)
+
+type tokens struct {
+	token                 string
+	accessToken           string
+	tokenExpiration       time.Time
+	accessTokenExpiration time.Time
+}
+
+type tokensAndError struct {
+	tokens *tokens
+	err    error
+}
+
+func (p *Provider) createTokens(ctx context.Context, saRef *serviceaccounts.Reference) (*tokens, string, error) {
+	now := time.Now()
+
+	sa, err := p.opts.ServiceAccounts.Get(ctx, saRef)
+	if err != nil {
+		return nil, "", fmt.Errorf("error getting kubernetes service account: %w", err)
+	}
+
+	email, err := serviceaccounts.GoogleEmail(sa)
+	if err != nil {
+		return nil, "", fmt.Errorf("error getting google service account from kubernetes service account: %w", err)
+	}
+
+	token, tokenDuration, err := p.opts.Source.GetServiceAccountToken(ctx, saRef)
+	if err != nil {
+		return nil, "", fmt.Errorf("error creating token for kubernetes service account: %w", err)
+	}
+
+	accessToken, accessTokenDuration, err := p.opts.Source.GetGoogleAccessToken(ctx, token, email)
+	if err != nil {
+		return nil, "", fmt.Errorf("error creating access token for google service account %s: %w", email, err)
+	}
+
+	return &tokens{
+		token:                 token,
+		accessToken:           accessToken,
+		tokenExpiration:       now.Add(tokenDuration),
+		accessTokenExpiration: now.Add(accessTokenDuration),
+	}, email, nil
+}
+
+func (t *tokens) sleepDurationUntilNextFetch() time.Duration {
+	sleepDuration := time.Until(t.tokenExpiration)
+	if d := time.Until(t.accessTokenExpiration); d < sleepDuration {
+		sleepDuration = d
+	}
+	const safeDistance = time.Minute
+	if sleepDuration >= safeDistance {
+		sleepDuration -= safeDistance
+	}
+	return sleepDuration
+}
